requirements.txt
google-api-python-client==2.70.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
python-dotenv==0.21.0
joblib==1.2.0
scikit-learn==1.2.2
pandas==1.5.3
